requirements.txt
argparse
PyYAML
lxml
httplib2
yamllint

pandas
numpy

matplotlib
calmap

oauth2client

pillow

progress
tqdm
termcolor
colorama

pysvg-py3
python-chess
